Implement DQN agent
- Implement DQN agent that learns how to play the game. Available
actions: call, check, raise by 50 (fixed), fold. Illegal actions
such as calling instead of checking are not forbidden, and instead
are corrected a posteriori. The observation space is the agent's
equity, its stack and the total pot value. Only one agent can exist
at a table.
- Update example script for one game, and add example training
script
- Change actions to action, i.e. only current player's action
intead of a table of useless values except that of the current
player